requirements.txt

fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-telegram-bot==20.8
agno
pydantic==2.6.4
httpx==0.27.0